requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
clickhouse-connect==0.4.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
celery==5.3.1

